+++ client/app/reader/CommentLayer.jsx
@@ -3,14 +3,20 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { makeGetAnnotationsByDocumentId } from '../reader/selectors';
 import CommentIcon from '../components/CommentIcon';
-import { keyOfAnnotation, pageNumberOfPageIndex, getPageCoordinatesOfMouseEvent } from './utils';
+import { keyOfAnnotation, pageNumberOfPageIndex, getPageCoordinatesOfMouseEvent,
+  isUserEditingText } from './utils';
 import _ from 'lodash';
-import { handleSelectCommentIcon, placeAnnotation } from '../reader/actions';
+import { handleSelectCommentIcon, placeAnnotation,
+  requestMoveAnnotation, startPlacingAnnotation,
+  stopPlacingAnnotation, showPlaceAnnotationIcon, } from '../reader/actions';
 import { bindActionCreators } from 'redux';
+import { CATEGORIES, INTERACTION_TYPES } from '../reader/analytics';
 
 const DIV_STYLING = {
   width: '100%',
-  height: '100%'
+  height: '100%',
+  zIndex: 10,
+  position: 'relative'
 };
 
 // The comment layer is a div on top of a page that draws the comment
@@ -44,8 +50,6 @@ class CommentLayer extends PureComponent {
     );
   };
 
-  getCommentLayerDivRef = (ref) => this.commentLayerDiv = ref
-
   getPlacingAnnotation = () => {
     if (this.props.placingAnnotationIconPageCoords && this.props.isPlacingAnnotation) {
       return [{
@@ -59,6 +63,121 @@ class CommentLayer extends PureComponent {
     return [];
 
   }
+  // Move the comment when it's dropped on a page
+  // eslint-disable-next-line max-statements
+  onCommentDrop = (event) => {
+    const dragAndDropPayload = event.dataTransfer.getData('text');
+    let dragAndDropData;
+
+    // Anything can be dragged and dropped. If the item that was
+    // dropped doesn't match what we expect, we just silently ignore it.
+    const logInvalidDragAndDrop = () => window.analyticsEvent(CATEGORIES.VIEW_DOCUMENT_PAGE, 'invalid-drag-and-drop');
+
+    try {
+      dragAndDropData = JSON.parse(dragAndDropPayload);
+
+      if (!dragAndDropData.iconCoordinates || !dragAndDropData.uuid) {
+        logInvalidDragAndDrop();
+
+        return;
+      }
+    } catch (err) {
+      if (err instanceof SyntaxError) {
+        logInvalidDragAndDrop();
+
+        return;
+      }
+      throw err;
+    }
+
+    const pageBox = this.commentLayerDiv.getBoundingClientRect();
+
+    const coordinates = {
+      x: (event.pageX - pageBox.left - dragAndDropData.iconCoordinates.x) / this.props.scale,
+      y: (event.pageY - pageBox.top - dragAndDropData.iconCoordinates.y) / this.props.scale
+    };
+
+    const droppedAnnotation = {
+      ...this.props.allAnnotations[dragAndDropData.uuid],
+      ...coordinates
+    };
+
+    this.props.requestMoveAnnotation(droppedAnnotation);
+  }
+
+  handleAltC = () => {
+    this.props.startPlacingAnnotation(INTERACTION_TYPES.KEYBOARD_SHORTCUT);
+
+    const scrollWindowBoundingRect = this.scrollWindow.getBoundingClientRect();
+    const firstPageWithRoomForIconIndex = pageIndexOfPageNumber(this.currentPage);
+
+    const iconPageBoundingBox =
+      this.pageElements[this.props.file][firstPageWithRoomForIconIndex].pageContainer.getBoundingClientRect();
+
+    const pageCoords = getInitialAnnotationIconPageCoords(
+      iconPageBoundingBox,
+      scrollWindowBoundingRect,
+      this.props.scale
+    );
+
+    this.props.showPlaceAnnotationIcon(firstPageWithRoomForIconIndex, pageCoords);
+  }
+
+  handleAltEnter = () => {
+    this.props.placeAnnotation(
+      pageNumberOfPageIndex(this.props.placingAnnotationIconPageCoords.pageIndex),
+      {
+        xPosition: this.props.placingAnnotationIconPageCoords.x,
+        yPosition: this.props.placingAnnotationIconPageCoords.y
+      },
+      this.props.documentId
+    );
+  }
+
+  keyListener = (event) => {
+    if (isUserEditingText()) {
+      return;
+    }
+
+    if (event.altKey) {
+      if (event.code === 'KeyC') {
+        this.handleAltC();
+      }
+
+      if (event.code === 'Enter') {
+        this.handleAltEnter();
+      }
+    }
+
+    if (event.code === 'Escape' && this.props.isPlacingAnnotation) {
+      this.props.stopPlacingAnnotation(INTERACTION_TYPES.KEYBOARD_SHORTCUT);
+    }
+  }
+
+  componentDidMount() {
+    window.addEventListener('keydown', this.keyListener);
+  }
+
+  componentWillUnmount() {
+    window.removeEventListener('keydown', this.keyListener);
+  }
+
+  mouseListener = (event) => {
+    if (this.props.isPlacingAnnotation) {
+      const pageCoords = getPageCoordinatesOfMouseEvent(
+        event,
+        this.commentLayerDiv.getBoundingClientRect(),
+        this.props.scale
+      );
+
+      this.props.showPlaceAnnotationIcon(this.props.pageIndex, pageCoords);
+    }
+  }
+
+  // To specify the component as droppable, we need to preventDefault on the event.
+  onPageDragOver = (event) => {event.preventDefault()}
+
+  getCommentLayerDivRef = (ref) => this.commentLayerDiv = ref
 
   getAnnotationsForPage = () => {
     return this.props.comments.concat(this.getPlacingAnnotation()).
@@ -78,7 +197,10 @@ class CommentLayer extends PureComponent {
     return <div
       id={`comment-layer-${this.props.pageIndex}`}
       style={DIV_STYLING}
+      onDragOver={this.onPageDragOver}
+      onDrop={this.onCommentDrop}
       onClick={this.onPageClick}
+      onMouseMove={this.mouseListener}
       ref={this.getCommentLayerDivRef}>
       {this.getCommentIcons()}
     </div>;
@@ -104,13 +226,18 @@ CommentLayer.propTypes = {
 const mapStateToProps = (state, ownProps) => ({
   ...state.readerReducer.ui.pdf,
   ..._.pick(state.readerReducer, 'placingAnnotationIconPageCoords'),
-  comments: makeGetAnnotationsByDocumentId(state.readerReducer)(ownProps.documentId)
+  comments: makeGetAnnotationsByDocumentId(state.readerReducer)(ownProps.documentId),
+  allAnnotations: state.readerReducer.annotations
 });
 
 const mapDispatchToProps = (dispatch) => ({
   ...bindActionCreators({
     placeAnnotation,
-    handleSelectCommentIcon
+    handleSelectCommentIcon,
+    requestMoveAnnotation,
+    startPlacingAnnotation,
+    stopPlacingAnnotation,
+    showPlaceAnnotationIcon
   }, dispatch)
 });
