+++ client/app/reader/constants.js
@@ -63,7 +63,6 @@ export const REQUEST_EDIT_ANNOTATION_FAILURE = 'REQUEST_EDIT_ANNOTATION_FAILURE'
 export const START_PLACING_ANNOTATION = 'START_PLACING_ANNOTATION';
 export const PLACE_ANNOTATION = 'PLACE_ANNOTATION';
 export const SHOW_PLACE_ANNOTATION_ICON = 'SHOW_PLACE_ANNOTATION_ICON';
-export const HIDE_PLACE_ANNOTATION_ICON = 'HIDE_PLACE_ANNOTATION_ICON';
 export const STOP_PLACING_ANNOTATION = 'STOP_PLACING_ANNOTATION';
 export const SELECT_ANNOTATION = 'SELECT_ANNOTATION';
 export const SET_PAGE_COORD_BOUNDS = 'SET_PAGE_COORD_BOUNDS';
